MaterialAlertDialog

+++ app/src/main/java/com/example/store/MainActivity.kt
@@ -1,3 +1,5 @@
+package com.example.stores_da
+import android.content.DialogInterface
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.recyclerview.widget.GridLayoutManager
@@ -98,11 +100,18 @@ doAsync {
 }
 }
 override fun onDeleteStore(storeEntity: StoreEntity) {
-    doAsync {
-        StoreApplication.database.storeDao().deleteStore(storeEntity)
-        uiThread {
-            mAdapter.delete(storeEntity)
-        }
-    }
+    MaterialAlertDialogBuilder(this)
+        .setTitle(R.string.dialog_delete_title)
+        .setPositiveButton(R.string.dialog_delete_confirm, { dialogInterface, i ->
+            doAsync {
+                StoreApplication.database.storeDao().deleteStore(storeEntity)
+                uiThread {
+                    mAdapter.delete(storeEntity)
+                }
+            }
+        })
+        .setNegativeButton(R.string.dialog_delete_cancel,null)
+        .show()
+
     mAdapter.update(storeEntity)
 }}
